requirements.txt
# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Web Search
tavily-python>=0.5.0

# Vector DB
qdrant-client>=1.12.0
sentence-transformers>=3.0.0

# API Server
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.10.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.24.0
